Sync QE to CE.

+++ install/addsontool.sh
@@ -13,6 +13,8 @@ then
  echo " addsonpatch.sh has to be run from the Quantum ESPRESSO root directory"
  echo "WHERE_SOURCE is the relative path to the sources of the Addson code "
  echo "WHERE_LINKS is the relative path to the QE directory where the addson sources have to be linked"
+ echo "at the moment it only allows for pure f90 routines to be linked in flib"
+ echo "or pure f90 modules to be linked in Modules"
  echo " -patch  : apply patch to Makefiles " 
  echo " -revert : revert Makefiles to original "
  echo " ]"
@@ -21,14 +23,26 @@ fi
 
 case "$4" in
 (-patch)
+
   echo "* I will try to patch needed files for integrated compilation ..."
+
+  if test -e "${ADDSON_NAME}_PATCH" ; then
+    echo "-- File $destination/${ADDSON_NAME}_PATCH exists"
+    echo "-- I guess you have already patched $ADDSON_NAME"
+    echo "-- Please unpatch it first, or start from a clean source tree"
+    echo "-- See you later..."
+    echo "* ABORT"
+    exit
+  fi
+  echo "#Please do not remove or modify this file"                    >  ${ADDSON_NAME}_PATCH
+  echo "#It is keeps track of the steps for patching $ADDSON package" >> ${ADDSON_NAME}_PATCH
   
 #-------------------
   echo "-- Executing pre script"
 
   command -v patch &>/dev/null || { echo "I require patch command but it's not installed. Aborting." >&2; exit 1;  }
 
-#------------------- first, check if GNU patch works
+#------------------- check if GNU patch works
   cat > test_patch1 << \EOF
   alfa
   beta
@@ -40,7 +54,7 @@ EOF
 EOF
 
   cat > test_patch3 << \EOF_EOF
-  patch -c -l -b -F 3 --suffix=.pre$ADDSON_NAME "./test_patch1" << \EOF
+  patch -c -l -b -F 3 --suffix=.pre "./test_patch1" << \EOF
 EOF_EOF
 
   diff -c test_patch1 test_patch2 >> test_patch3
@@ -61,13 +75,51 @@ EOF_EOF
   fi
 
   rm test_patch1 test_patch2 test_patch3 test_patch4
-  if [ -e test_patch1.pre$ADDSON_NAME ]
+  if [ -e test_patch1.pre ]
   then
-    rm test_patch1.pre$ADDSON_NAME
+    rm test_patch1.pre
   fi
-#-------------------
+#-------------------------------------------
 
+  command -v sed &>/dev/null || { echo "I require sed command but it's not installed. Aborting." >&2; exit 1;  }
 
+#------------------- check if GNU sed works
+  cat > test_sed1 << \EOF
+  alfa
+  beta
+EOF
+
+  cat > test_sed2 << \EOF
+  alfa
+  gamma
+  beta
+EOF
+
+  sed '/alfa/ a\
+  gamma' test_sed1 > tmp.1
+
+  mv tmp.1 test_sed1
+
+  diff -c test_sed1 test_sed2 >> test_sed3
+
+#  echo EOF >> test_sed3
+
+  bash test_sed3 &> test_sed4
+
+  status=$?
+  if [ $status -ne 0 ]
+  then
+    echo "sed does not work! Error message:"
+    echo "**********"
+    cat test_sed4
+    echo "**********"
+    echo "Please install a recent version of the GNU sed utility and try again."
+    exit
+  fi
+
+  rm test_sed1 test_sed2 test_sed3 test_sed4
+# -----------------------------------------
+# -----------------------------------------
   to_do_before_patch
 
   echo "-- Setting up symlinks"
@@ -125,7 +177,8 @@ EOF_EOF
     echo "* ABORT"
     exit
   fi
-
+  
+  rm ${ADDSON_NAME}_PATCH
 
   echo "-- Executing post script"
   to_do_after_revert
